refactor(source): share client-side search across custom source types
